iox-#25: moved reused ports to ports folder
