+++ iceoryx_posh/include/iceoryx_posh/internal/popo/ports/base_port.inl
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iceoryx_posh/internal/popo/base_port.hpp"
+#include "iceoryx_posh/internal/popo/ports/base_port.hpp"
 
 namespace iox
 {
