+++ iceoryx_posh/include/iceoryx_posh/internal/popo/receiver_port_data.hpp
@@ -14,8 +14,7 @@
 
 #pragma once
 
-#include "iceoryx_posh/internal/popo/base_port_data.hpp"
-
+#include "iceoryx_posh/internal/popo/ports/base_port_data.hpp"
 #include "iceoryx_posh/capro/service_description.hpp"
 #include "iceoryx_posh/internal/popo/delivery_fifo.hpp"
 #include "iceoryx_posh/internal/popo/used_chunk_list.hpp"
