+++ iceoryx_posh/include/iceoryx_posh/internal/popo/sender_port.hpp
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "iceoryx_posh/internal/popo/base_port.hpp"
+#include "iceoryx_posh/internal/popo/ports/base_port.hpp"
 #include "iceoryx_posh/internal/popo/sender_port_data.hpp"
 
 #include <cstdint>
