+++ iceoryx_posh/include/iceoryx_posh/popo/gateway_generic.hpp
@@ -16,7 +16,7 @@
 
 #include "iceoryx_posh/capro/service_description.hpp"
 #include "iceoryx_posh/iceoryx_posh_types.hpp"
-#include "iceoryx_posh/internal/popo/interface_port.hpp"
+#include "iceoryx_posh/internal/popo/ports/interface_port.hpp"
 
 #include <memory>
