+++ iceoryx_posh/source/popo/ports/interface_port.cpp
@@ -12,23 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iceoryx_posh/internal/popo/application_port.hpp"
+#include "iceoryx_posh/internal/popo/ports/interface_port.hpp"
+#include "iceoryx_utils/error_handling/error_handling.hpp"
 
 namespace iox
 {
 namespace popo
 {
-ApplicationPort::ApplicationPort(ApplicationPortData* const f_memberPtr)
-    : BasePort(f_memberPtr)
+InterfacePort::InterfacePort(InterfacePortData* const f_member)
+    : BasePort(f_member)
 {
 }
 
-bool ApplicationPort::dispatchCaProMessage(const capro::CaproMessage& f_message)
+bool InterfacePort::dispatchCaProMessage(const capro::CaproMessage& f_message)
 {
-    return getMembers()->m_caproMessageFiFo.push(f_message);
+    bool returner = getMembers()->m_caproMessageFiFo.push(f_message);
+    if (!returner)
+    {
+        errorHandler(Error::kPOSH__INTERFACEPORT_CAPRO_MESSAGE_DISMISSED, nullptr, iox::ErrorLevel::SEVERE);
+    }
+    return returner;
 }
 
-bool ApplicationPort::getCaProMessage(capro::CaproMessage& f_message)
+bool InterfacePort::getCaProMessage(capro::CaproMessage& f_message)
 {
     auto msg = getMembers()->m_caproMessageFiFo.pop();
     if (msg.has_value())
@@ -39,15 +45,16 @@ bool ApplicationPort::getCaProMessage(capro::CaproMessage& f_message)
     return false;
 }
 
-const typename ApplicationPort::MemberType_t* ApplicationPort::getMembers() const
+const InterfacePortData* InterfacePort::getMembers() const
 {
-    return reinterpret_cast<const MemberType_t*>(BasePort::getMembers());
+    return reinterpret_cast<const InterfacePortData*>(BasePort::getMembers());
 }
 
-typename ApplicationPort::MemberType_t* ApplicationPort::getMembers()
+InterfacePortData* InterfacePort::getMembers()
 {
-    return reinterpret_cast<MemberType_t*>(BasePort::getMembers());
+    return reinterpret_cast<InterfacePortData*>(BasePort::getMembers());
 }
+
 
 } // namespace popo
 } // namespace iox
