+++ iceoryx_posh/source/popo/ports/interface_port_data.cpp
@@ -10,26 +10,20 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License
+// limitations under the License.
 
-#include "iceoryx_posh/internal/popo/base_port_data.hpp"
+#include "iceoryx_posh/internal/popo/ports/interface_port_data.hpp"
 
 namespace iox
 {
 namespace popo
 {
-BasePortData::BasePortData() noexcept
-    : m_uniqueId(s_uniqueIdCounter.fetch_add(1u, std::memory_order_relaxed))
-{
-}
-
-BasePortData::BasePortData(const capro::ServiceDescription& serviceDescription,
-                           const BasePortType& portType,
-                           const cxx::CString100& processName) noexcept
-    : m_portType(portType)
-    , m_serviceDescription(serviceDescription)
-    , m_processName(processName)
-    , m_uniqueId(s_uniqueIdCounter.fetch_add(1u, std::memory_order_relaxed))
+InterfacePortData::InterfacePortData(const std::string& applicationName,
+                                     const capro::Interfaces interface) noexcept
+    : BasePortData(capro::ServiceDescription(
+                       capro::InvalidIDString, capro::InvalidIDString, capro::InvalidIDString, {0, 0, 0, 0}, interface),
+                   BasePortType::INTERFACE_PORT,
+                   iox::cxx::string<100>(iox::cxx::TruncateToCapacity, applicationName))
 {
 }
 } // namespace popo
